Gini coefficient

The Gini coefficient is a single number aimed at measuring the degree of inequality in a distribution. It is usually defined mathematically based on the Lorenz curve. A Gini coefficient of zero expresses perfect equality, where all values are the same, whereas a coefficient of 1 expresses maximal inequality among values.

CalculatedABC.gini_coeff — Method

gini_coeff(data::Vector{<:Real})

Calculates the Gini coefficient for the given data.

CalculatedABC.gini_coeff — Method

gini_coeff(curve::ABCcurve)

Calculates the Gini coefficient for the given ABC curve.