ABC Analysis

CalculatedABC.ABCanalysis — Type.

ABCanalysis(curve::ABCcurve)

Calculate an ABC analysis for the given curve.

Fields

pareto*point: Point on the curve at which most of the yield is already obtained. Demarkation between A and B.

break*even*point: Point on the curve where the gain ($dABC$) is 1.

submarginal*point: Point on the curve after which the gain can be considered trivial. Demarkation between B and C.

source